+++ config/routes.rb
@@ -145,6 +145,7 @@ Rails.application.routes.draw do
         get "events/:event_id/:subject_event_id/:event_date/edit", action: :edit_event, as: :edit_event
         post "events/:event_id/:subject_event_id/:event_date", action: :update_event, as: :update_event
         get :timeline
+        get :comments
         get :settings
         get :files
         get :sheets